functions/pubsub-negative-analyzer/requirements.txt
Flask
google-cloud-pubsub
google-cloud-language
slack_sdk